+++ src/main/java/com/sparta/spring_deep/_delivery/admin/controller/Controller.java
@@ -1,0 +1,5 @@
+package com.sparta.spring_deep._delivery.admin.controller;
+
+public class Controller {
+
+}
