+++ src/main/java/com/sparta/spring_deep/_delivery/admin/dto/Dto.java
@@ -1,0 +1,5 @@
+package com.sparta.spring_deep._delivery.admin.dto;
+
+public class Dto {
+
+}
