+++ src/main/java/com/sparta/spring_deep/_delivery/admin/repository/AiRepository.java
@@ -1,5 +1,6 @@
-package com.sparta.spring_deep._delivery.domain.ai;
+package com.sparta.spring_deep._delivery.admin.repository;
 
+import com.sparta.spring_deep._delivery.domain.ai.Ai;
 import java.util.UUID;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
