+++ src/main/java/com/sparta/spring_deep/_delivery/admin/repository/Repository.java
@@ -1,0 +1,5 @@
+package com.sparta.spring_deep._delivery.admin.repository;
+
+public interface Repository {
+
+}
