+++ src/main/java/com/sparta/spring_deep/_delivery/admin/service/Service.java
@@ -1,0 +1,5 @@
+package com.sparta.spring_deep._delivery.admin.service;
+
+public class Service {
+
+}
